rename: 공용 컴포넌트 폴더 이동
