+++ src/components/ProductItem/index.tsx
@@ -1,7 +1,7 @@
 import useProductSelect from 'src/hooks/useCartUpdate';
 import { Product } from 'src/types';
 import Svg from '../@common/Svg';
-import Counter from '../Counter';
+import Counter from '../@common/Counter';
 import * as S from './ProductItem.styles';
 import { convertKORWon } from 'src/utils';
